refactor(TextInputInnerAction): moved to internal
